+++ docs/src/examples/framework/maximal/component.js
@@ -19,6 +19,7 @@ import {
   onAttributeChanged,
   onCreated,
   onDestroyed,
+  onRendered,
   onThemeChanged
 } from './lifecycle.js';
 
@@ -29,6 +30,7 @@ defineComponent({
   subTemplates: { buttons },
   createComponent,
   onCreated,
+  onRendered,
   onDestroyed,
   onThemeChanged,
   onAttributeChanged,