+++ examples/test-case/index.js
@@ -1,3 +1,0 @@
-import { TestCase } from './test-case.js';
-
-export { TestCase };